+++ AGENTS.md
@@ -53,7 +53,7 @@ Release-oriented notes also live in [`docs/RELEASE.md`](docs/RELEASE.md) under p
 
 `agent_browser` results include stable machine-readable fields on `details`: `resultCategory` (`success` | `failure`), plus `successCategory` or `failureCategory` with small fixed enums. The human-facing contract and field list live in [`docs/TOOL_CONTRACT.md`](docs/TOOL_CONTRACT.md#details).
 
-- **Module map (maintainers):** `extensions/agent-browser/lib/session-page-state.ts` owns ordered per-session tab target, `refSnapshot` / invalidation, and pinning state driven from `extensions/agent-browser/index.ts` (see `test/agent-browser.session-page-state.test.ts`). Top-level structured modes live under `extensions/agent-browser/lib/input-modes/` (`params.ts`, `job.ts`, `semantic-action.ts`, `lookups.ts`, `electron.ts`, and `types.ts`), with `lib/input-modes.ts` as a re-export-only facade. Input resolution/validation lives in `extensions/agent-browser/lib/orchestration/input-plan.ts`; browser command execution phases live in `extensions/agent-browser/lib/orchestration/browser-run/`, with `browser-run.ts` as a compatibility re-export; keep future execution-phase helpers under `lib/orchestration/` rather than growing `index.ts`. Under `extensions/agent-browser/lib/results/`, `shared.ts` is a **re-export-only** barrel; edit behavior in the focused file (`contracts.ts`, `categories.ts`, `action-recommendations.ts`, `next-actions.ts`, `recovery-actions.ts`, `recovery-next-actions.ts`, `selector-recovery.ts`, `network.ts`, `artifact-manifest.ts`, `artifact-state.ts`, `snapshot.ts` plus `snapshot-*` helpers, `editable-ref-evidence.ts`, `text.ts`, `presentation.ts`, and `presentation/*` helpers) rather than adding logic back into `shared.ts`.
+- **Module map (maintainers):** `extensions/agent-browser/lib/session-page-state.ts` owns ordered per-session tab target, `refSnapshot` / invalidation, and pinning state driven from `extensions/agent-browser/index.ts` (see `test/agent-browser.session-page-state.test.ts`). Top-level structured modes live under `extensions/agent-browser/lib/input-modes/` (`params.ts`, `job.ts`, `semantic-action.ts`, `lookups.ts`, `electron.ts`, and `types.ts`), with `lib/input-modes.ts` as a re-export-only facade. Input resolution/validation lives in `extensions/agent-browser/lib/orchestration/input-plan.ts`; browser command execution phases live in `extensions/agent-browser/lib/orchestration/browser-run/`, with `browser-run.ts` as a compatibility re-export; wrapper-side Electron host actions live in `extensions/agent-browser/lib/orchestration/electron-host/`. Keep future execution-phase helpers under `lib/orchestration/` rather than growing `index.ts`. Under `extensions/agent-browser/lib/results/`, `shared.ts` is a **re-export-only** barrel; edit behavior in the focused file (`contracts.ts`, `categories.ts`, `action-recommendations.ts`, `next-actions.ts`, `recovery-actions.ts`, `recovery-next-actions.ts`, `selector-recovery.ts`, `network.ts`, `artifact-manifest.ts`, `artifact-state.ts`, `snapshot.ts` plus `snapshot-*` helpers, `editable-ref-evidence.ts`, `text.ts`, `presentation.ts`, and `presentation/*` helpers) rather than adding logic back into `shared.ts`.
 - **Pi transcript alignment:** Pi’s custom-tool path normally marks a tool row failed only when `execute` throws, so a wrapper that returns success while setting `details.resultCategory: "failure"` (for example after `qa` diagnostic reclassification) would otherwise look like a passing tool in JSONL/UI. The extension registers `pi.on("tool_result", …)` and `buildAgentBrowserToolResultPatch` in `extensions/agent-browser/index.ts` patches matching `agent_browser` results to set `isError: true` when `event.isError` was not already true. For prose results it also appends the visible `Result category: failure; …; Pi tool isError: true.` line; for caller-requested `--json` results it must preserve the parseable JSON string and patch only `isError`. When changing that behavior, update [`docs/TOOL_CONTRACT.md`](docs/TOOL_CONTRACT.md#details), [`docs/ARCHITECTURE.md`](docs/ARCHITECTURE.md), release notes if transcript review guidance shifts, and extension-validation expectations that assert both prose notice text and JSON preservation.
 - **Source of truth for enums and classifiers:** `extensions/agent-browser/lib/results/categories.ts` (`classifyAgentBrowserSuccessCategory`, `classifyAgentBrowserFailureCategory`, `buildAgentBrowserResultCategoryDetails`), re-exported from `extensions/agent-browser/lib/results/shared.ts` for compatibility. The tool entrypoint merges category details in `extensions/agent-browser/index.ts`; presentation-layer failures also attach categories from `extensions/agent-browser/lib/results/presentation.ts`.
 - **Structured follow-ups:** optional `details.nextActions` (and per-step `batchSteps[].nextActions`) are built by `buildAgentBrowserNextActions` in `extensions/agent-browser/lib/results/action-recommendations.ts` (re-exported from `shared.ts`), then merged into Pi-facing `details` from `extensions/agent-browser/index.ts` and `extensions/agent-browser/lib/results/presentation.ts`. See [`docs/TOOL_CONTRACT.md`](docs/TOOL_CONTRACT.md#details) for the field contract.
@@ -83,7 +83,7 @@ Release-oriented notes also live in [`docs/RELEASE.md`](docs/RELEASE.md) under p
 ### Electron desktop lifecycle (top-level `electron`)
 
 - **Host modules:** `extensions/agent-browser/lib/electron/discovery.ts` (install scans and Electron evidence; `discoverElectronApps` scans macOS/Linux only—Windows returns `platform: "unsupported"` for `list`, while `inspectElectronExecutablePath` can still verify Windows binaries for `launch`), `launch.ts` (isolated `userDataDir`, `--remote-debugging-port=0`, `DevToolsActivePort` polling, CDP metadata), `cleanup.ts` (tracked PID/port/session/profile teardown and status probes).
-- **Extension entrypoint:** `extensions/agent-browser/lib/input-modes/electron.ts` owns `compileAgentBrowserElectron`; `extensions/agent-browser/index.ts` owns transcript restore for wrapper `launchId` records (`restoreElectronLaunchRecordsFromBranch`), orchestration into upstream `connect` / handoff batches, merged `details.electron*` / `details.compiledElectron`, post-command health checks, and structured next actions documented in [`docs/TOOL_CONTRACT.md`](docs/TOOL_CONTRACT.md#electron).
+- **Orchestration modules:** `extensions/agent-browser/lib/input-modes/electron.ts` owns `compileAgentBrowserElectron`; `extensions/agent-browser/lib/orchestration/electron-host/index.ts` owns wrapper host actions (`list`, `status`, `probe`, `cleanup`), transcript restore for wrapper `launchId` records (`restoreElectronLaunchRecordsFromBranch`), and host-action result assembly. `extensions/agent-browser/lib/orchestration/browser-run/prepare.ts` / `final-result.ts` own `electron.launch` connect and handoff execution plus merged `details.electron*` / `details.compiledElectron`, post-command health checks, and structured next actions documented in [`docs/TOOL_CONTRACT.md`](docs/TOOL_CONTRACT.md#electron). `extensions/agent-browser/index.ts` should stay limited to registration, session state initialization, cleanup dispatch, and tool-call routing.
 - **When changing behavior:** update [`docs/TOOL_CONTRACT.md`](docs/TOOL_CONTRACT.md#electron), [`docs/ELECTRON.md`](docs/ELECTRON.md) (public lifecycle and troubleshooting guide), [`docs/COMMAND_REFERENCE.md`](docs/COMMAND_REFERENCE.md), [`README.md`](README.md) agent copy, [`docs/SUPPORT_MATRIX.md`](docs/SUPPORT_MATRIX.md) `RQ-0096`, playbook strings in `extensions/agent-browser/lib/playbook.ts` when operator guidance shifts (`npm run docs -- playbook write`), plus `test/agent-browser.extension-validation.test.ts`, `test/agent-browser.results.test.ts`, and benchmark wiring in `scripts/agent-browser-efficiency-benchmark.mjs` with `test/agent-browser.efficiency-benchmark.test.ts`.
 
 ### Constrained `job` (top-level tool input)
